Reorganise Server File Structure
