+++ server/src/models/User.ts
@@ -6,6 +6,7 @@ interface IUser {
   name: string;
   email: string;
   password: string;
+  profileImage?: string;
 }
 
 const UserSchema = new Schema<IUser>({
